Add a static component paginator + Support for localising paginated Pages
