+++ yuyo/__init__.py
@@ -93,9 +93,10 @@ from .components import ComponentClient
 from .components import ComponentContext
 from .components import ComponentExecutor
 from .components import ComponentPaginator
-from .components import InteractionError
+
 from .components import StreamExecutor
 from .components import WaitForExecutor
+from .interactions import InteractionError
 from .list_status import BotsGGService
 from .list_status import DiscordBotListService
 from .list_status import ServiceManager
